Moved Autocomplete methods into their correct parent classes
