+++ src/main/java/net/javadiscord/javabot/systems/qotw/subcommands/questions_queue/ListQuestionsSubcommand.java
@@ -1,11 +1,8 @@
 package net.javadiscord.javabot.systems.qotw.subcommands.questions_queue;
 
 import net.dv8tion.jda.api.EmbedBuilder;
-import net.dv8tion.jda.api.events.interaction.command.CommandAutoCompleteInteractionEvent;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
-import net.dv8tion.jda.api.interactions.commands.Command;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
-import net.dv8tion.jda.api.requests.restaction.interactions.AutoCompleteCallbackAction;
 import net.dv8tion.jda.api.requests.restaction.interactions.ReplyCallbackAction;
 import net.javadiscord.javabot.Bot;
 import net.javadiscord.javabot.command.Responses;
@@ -16,8 +13,6 @@ import net.javadiscord.javabot.systems.qotw.subcommands.QOTWSubcommand;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.time.ZoneOffset;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Subcommand that allows staff-members to list QOTW Questions.
@@ -61,23 +56,5 @@ public class ListQuestionsSubcommand extends QOTWSubcommand {
 			event.getHook().sendMessageEmbeds(embedBuilder.build()).queue();
 		});
 		return event.deferReply();
-	}
-
-	/**
-	 * Replies with all Question of the Week Questions.
-	 *
-	 * @param event The {@link CommandAutoCompleteInteractionEvent} that was fired.
-	 * @return The {@link AutoCompleteCallbackAction}.
-	 */
-	public static AutoCompleteCallbackAction replyQuestions(CommandAutoCompleteInteractionEvent event) {
-		List<Command.Choice> choices = new ArrayList<>(25);
-		try (Connection con = Bot.dataSource.getConnection()) {
-			QuestionQueueRepository repo = new QuestionQueueRepository(con);
-			List<QOTWQuestion> questions = repo.getQuestions(event.getGuild().getIdLong(), 0, 25);
-			questions.forEach(question -> choices.add(new Command.Choice(String.format("(Priority: %s) %s", question.getPriority(), question.getText()), question.getId())));
-		} catch (SQLException e) {
-			e.printStackTrace();
-		}
-		return event.replyChoices(choices);
 	}
 }
